+++ security/source.go
@@ -1,8 +1,6 @@
 package security
 
 import (
-	"strings"
-
 	"github.com/imgproxy/imgproxy/v2/config"
 )
 
@@ -10,8 +8,8 @@ func VerifySourceURL(imageURL string) bool {
 	if len(config.AllowedSources) == 0 {
 		return true
 	}
-	for _, val := range config.AllowedSources {
-		if strings.HasPrefix(imageURL, string(val)) {
+	for _, allowedSource := range config.AllowedSources {
+		if allowedSource.MatchString(imageURL) {
 			return true
 		}
 	}